### 2024-02-04 #6320 @danielskatz
Note from author when submitting:
> The original version of this software (v1.0.0) was published in Environmental Modelling & Software (Nguyen et al., 2022). In the submitted version to JOSS (v4.0.0) I have made sustainability changes (please see the manuscipt)
