### 2024-02-04 #6320 @danielskatz
👋 @tamnva - I assume the paper is in the joss_manuscript branch?  If so, please add the software there too. JOSS wants the paper and the software to be in a branch, which could be main but can also be another branch.
