### 2024-02-04 #6320 @danielskatz
@tamnva - Also, JOSS is unhappy if the name of the branch with the paper is the same as the root name of the paper itself, so can you change the name of the files that contain the paper and bib, perhaps paper.md and paper.bib? (and you will, of course, have to change the reference to the bib file in the md file as well)
