### 2024-02-04 52b99bf
change paper and bib name, remove unused references
