+++ paper.bib
@@ -37,20 +37,6 @@
   publisher={American Society of Agricultural and Biological Engineers}
 }
 
-@article{nguyen2018,
-	author = {Nguyen, Van Tam and Dietrich, J{\"o}rg},
-	title = {Modification of the SWAT model to simulate regional groundwater flow using a multicell aquifer},
-	journal = {Hydrological Processes},
-	volume = {32},
-	number = {7},
-	pages = {939-953},
-	keywords = {baseflow, groundwater flow, hydrological connectivity, multicell aquifer, SWAT-MCA},
-	doi = {https://doi.org/10.1002/hyp.11466},
-	url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/hyp.11466},
-	eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/hyp.11466},
-	year = {2018}
-}
-
 @article{aloui2023,
 	title = {A review of Soil and Water Assessment Tool (SWAT) studies of Mediterranean catchments: Applications, feasibility, and future directions},
 	journal = {Journal of Environmental Management},
