### 2024-02-04 #6320 @danielskatz
@tamnva - Given that there was a previous publication, I am going to ask other JOSS editors to check this to make sure the updates are of sufficient scope for JOSS to review this.  This should take a week or two.

I believe the diffs can be seen via https://github.com/tamnva/R-SWAT/compare/v1.0.0...v4.0.0
