+++ joss/paper.bib
@@ -6,7 +6,7 @@
 	number = {1},
 	pages = {115-130},
 	keywords = {watershed management, computational methods, simulation, watersheds, rivers/streams},
-	doi = {https://doi.org/10.1111/1752-1688.12482},
+	doi = {10.1111/1752-1688.12482},
 	url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/1752-1688.12482},
 	eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/1752-1688.12482},
 	year = {2017}
@@ -20,7 +20,7 @@
 	number = {1},
 	pages = {73-89},
 	keywords = {simulation, surface water hydrology, erosion, sedimentation, nonpoint source pollution, large area modeling, plant growth, agricultural land management},
-	doi = {https://doi.org/10.1111/j.1752-1688.1998.tb05961.x},
+	doi = {10.1111/j.1752-1688.1998.tb05961.x},
 	url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1752-1688.1998.tb05961.x},
 	eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/j.1752-1688.1998.tb05961.x},
 	year = {1998}
@@ -33,6 +33,7 @@
   volume={50},
   number={4},
   pages={1211--1250},
+  doi = {10.13031/2013.23637},
   year={2007},
   publisher={American Society of Agricultural and Biological Engineers}
 }
@@ -44,7 +45,7 @@
 	pages = {116799},
 	year = {2023},
 	issn = {0301-4797},
-	doi = {https://doi.org/10.1016/j.jenvman.2022.116799},
+	doi = {10.1016/j.jenvman.2022.116799},
 	url = {https://www.sciencedirect.com/science/article/pii/S0301479722023726},
 	author = {Sarra Aloui and Annamaria Mazzoni and Adel Elomri and Jalel Aouissi and Abdelmadjid Boufekane and Adel Zghibi},
 	keywords = {Soil and water assessment tool, Watershed modeling, Mediterranean catchment, Model performance},
@@ -58,7 +59,7 @@
   number = {5},
   pages = {3941-3959},
   keywords = {Nitrogen, Legacy, Lag Time, SWAT, Travel Time Distribution, Best Management Practices},
-  doi = {https://doi.org/10.1029/2018WR023815},
+  doi = {10.1029/2018WR023815},
   url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2018WR023815},
   eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2018WR023815},
   year = {2019}
@@ -71,7 +72,7 @@
   pages = {105497},
   year = {2022},
   issn = {1364-8152},
-  doi = {https://doi.org/10.1016/j.envsoft.2022.105497},
+  doi = {10.1016/j.envsoft.2022.105497},
   url = {https://www.sciencedirect.com/science/article/pii/S1364815222002006},
   author = {Tam Van Nguyen and J{\"o}g Dietrich and Thanh Duc Dang and Dang An Tran and Binh {Van Doan} and Fanny J. Sarrazin and Karim Abbaspour and Raghavan Srinivasan}
 }
@@ -109,7 +110,7 @@
   pages = {138575},
   year = {2023},
   issn = {0959-6526},
-  doi = {https://doi.org/10.1016/j.jclepro.2023.138575},
+  doi = {10.1016/j.jclepro.2023.138575},
   url = {https://www.sciencedirect.com/science/article/pii/S0959652623027336},
[…]
