### 2024-02-04 #6320 @tamnva
> 'MISSING' DOIs are suggestions
invalid DOIs were fixed
