### 2024-02-04 #6320 @tamnva
> 👋 @tamnva - now, in addition to my previous request, to merge the code into the joss_manuscript branch, you could also work on the possibly missing DOI that editorialbot suggests, but note that some may be incorrect, and the DOIs that have the prefix 'https://doi.org/'. Please make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
