### 2024-02-04 #6320 @danielskatz
> editorialbot commands need to be the first entry in a new comment.
