### 2024-02-05 #6320 @tamnva
> @tamnva - Given that there was a previous publication, I am going to ask other JOSS editors to check this to make sure the updates are of sufficient scope for JOSS to review this. This should take a week or two.
>
> I believe the diffs can be seen via [tamnva/R-SWAT@v1.0.0...v4.0.0](https://github.com/tamnva/R-SWAT/compare/v1.0.0...v4.0.0)

Please let me know once the JOSS editor's decision is made, so I can continue and suggest potential reviewers. Thanks
