### 2024-02-05 ee39772
Update README.md
create R-CMD check status
