### 2024-02-05 e7edcf7
add R-CMD check git readme
