+++ man/calObjFunction.Rd
@@ -41,6 +41,9 @@ calibration and validation period and for each variable
 }
 \examples{
 
-# Please see RSWAT vignettes "SUFI2_without_GUI"
+\donttest{
+# Please see RSWAT Vignettes
+vignette("SUFI2_without_GUI",package="RSWAT")
+}
 
 }
