+++ NAMESPACE
@@ -19,6 +19,7 @@ export(lhsRange)
 export(loadSwatParam)
 export(perCriteria)
 export(prFactor)
+export(readAllOutVar)
 export(readChannelFile)
 export(readOutputRchFile)
 export(readWatoutFile)
