### 2024-02-08 ca9d9a5
generate html vignettes
