+++ vignettes/run_SWATplus_with_given_parametersets.R
@@ -1,0 +1,134 @@
+## ---- include = FALSE---------------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+
+## ---- message=FALSE, eval = FALSE---------------------------------------------
+#  library(RSWAT)
+#  
+#  # 1.Create an example TxtInOut of SWAT+ (TODO: remove this command if you already # had the TxtInOut folder of SWAT+)
+#  extracExampleData(exampleData, "swatPlusTxtInOut", tempdir())
+#  
+#  # 2.Path to the TxtInOut folder (TODO: replace with your TxtInOut path)
+#  TxtInOutFolder <- file.path(tempdir(), "swatPlusTxtInOut")
+#  
+#  # 3.This is a SWAT+ project, therefore
+#  SWATproject <- FALSE
+#  
+#  # 4.My working directory (TODO: change to your desired working directory)
+#  # dir.create(file.path(tempdir(), "workingFolder"))
+#  workingFolder <- file.path(tempdir(), "workingFolder")
+#  
+#  # 5.Path of the SWAT executable file (TODO: replace with your path)
+#  SWATexeFile <- "C:/data/rev60.5.7_64rel.exe"
+#  
+#  # 6.Path of the SWAT+ parameter file cal_parms.cal (TODO: replace with your path)
+#  SWATParamFile <- file.path(TxtInOutFolder, "cal_parms.cal")
+#  
+#  # 7.The parameters that I previously used for model calibration (TODO: change to your paraSelection)
+#  paraSelection <- data.frame(
+#     Parameter = c("cn2.hru"   , "canmx.hru"),
+#     Change =    c("relative"  , "replace"),
+#     Min =       c(-0.25       , 1.0),
+#     Max =       c(0.25        , 10.0),
+#     Object  =   c("All"       , "All"),
+#     Conditions = c("All"      , "All")
+#     )
+#  
+#  # You can also create these above table by the interface of R-SWAT then save
+#  # the  project. If you do so, you can extract that table from the RSWATproject.rds
+#  # file in the working folder (TODO: Delete if you don't read from RSWATproject.rds)
+#  
+#  # RSWATproject <- readRDS("C:/data/RSWATproject.rds")
+#  # paraSelection <- RSWATproject$paraSelection
+#  
+#  # 8. How parameters should be generated
+#  samplingApproach <- "Read_User_Parameter_File"  # Do not change this line
+#  
+#  # 9. Output extraction for SWAT+ can be created manually or by the GUI (TODO: change this)
+#  outputExtraction <- data.frame(
+#     FileType = c("channel_sd_day.txt"),
+#     FileName = c("channel_sd_day.txt"),
+#     Column = c("48"),
+#     Reach = c("1")
+#   )
+#  
+#  # 10. Date range for extraction (TODO: change this)
+#  dateRangeCali <- as.Date(c("2003-01-01", "2012-12-31"), format = "%Y-%m-%d")
+#  
+#  # 11. Number of parallel runs  (TODO: change  this)
+#  ncores <- 2
+#  
+#  # 12.  Generate parameter samples using Latin Hypercube Sampling (TODO: change this)
+#  parameterValue <- as.matrix(read.table(file = "C:/data/user_parameter_values.txt",
+#                                         header = TRUE, sep =""))
+#  
+#  # Here is the content of my user_parameter_values.txt"
+#  #cn2.hru         canmax.hru
+#  #0.10            1.5
+#  #-0.20           3.0
+#  #-0.10           0.2
+#  #0.22            1.0
+#  #
+#  
[…]
