### 2024-02-08 #6320 @danielskatz
@tamnva - We'll go ahead and review this.  However, note that one of the editors commented

>This isn't really packaged to normal R standards (it seems like users are intended to clone the repo and edit scripts before installing the package, which isn't typical) and there's a lot of odd code choices (long chain of "x == y | x == z |", rather than using %in%; bulky wrappers around built-in functions).
