### 2024-02-08 #6320 @tamnva
Please find below the list of potential reviewers which had R packages that are relevant to this work:

cvitolo
Sibada
ldecicco-USGS
odelaigue
jmp75

There is also a list of people that starred this GitHub repository (please see the link below), I am not sure whether I am allowed to suggest some of them or not

https://github.com/tamnva/R-SWAT/stargazers
