### 2024-02-08 #6320 @danielskatz
> There is also a list of people that starred this GitHub repository (please see the link below), I am not sure whether I am allowed to suggest some of them or not

Yes, that's fine to do as long as they are not contributors or otherwise conflicted. Being a user is not a conflict and can be a plus.
