### 2024-02-09 ca7ea8a
fix: error in readAllOutVar causes program to stop
