+++ vignettes/run_SWATplus_user_parameters.Rmd
@@ -75,8 +75,12 @@ outputExtraction <- data.frame(
    FileType = c("channel_sd_day.txt"),
    FileName = c("channel_sd_day.txt"),
    Column = c("48"),
-   Reach = c("1")
+   Reach = c("1") # E.g., for 100 reaches use: Reach = paste(as.character(c(1:100)), collapse  = ",")
  )
+# NOTE: if you want to extract for many reaches (e.g., for the first 100 reaches), create variable for th
+
+reach <- paste(as.character(c(1:100)), collapse  = ",")
+
 
 # 10. Date range for extraction (TODO: change this)
 dateRangeCali <- as.Date(c("2003-01-01", "2012-12-31"), format = "%Y-%m-%d")
