### 2024-02-09 96b4c70
clean vignettes
