### 2024-02-09 tamnva/R-SWAT#67 @tamnva
joss editor
Comments from the JOSS editor (https://github.com/openjournals/joss-reviews/issues/6320)

This isn't really packaged to normal R standards (it seems like users are intended to clone the repo and edit scripts before installing the package, which isn't typical) and there's a lot of odd code choices (long chain of "x == y | x == z |", rather than using %in%; bulky wrappers around built-in functions).
