+++ R/readWatoutFile.R
@@ -18,8 +18,7 @@ readWatoutFile <- function(workingDirectory,
                            fileCioInfo,
                            output){
 
-  fileName <- paste(workingDirectory, "/TxtInOut_",
-                    coreNumber, "/", fileName, sep = "")
+  fileName <- file.path(workingDirectory, "TxtInOut_", coreNumber, fileName)
 
   getWatoutData <- read.table(fileName, header = FALSE, sep = "", skip = 6)
 
