### 2024-02-09 tamnva/R-SWAT#67 @tamnva
Comment: users are intended to clone the repo and edit scripts before installing the package
Response: There is no need to edit scripts before installing the package now.

Comment: there's a lot of odd code choices (long chain of "x == y | x == z |", rather than using %in%
Response: The updated version uses %in%

Comment: bulky wrappers around built-in functions
Response: Working to split complex functions to smaller functions (TODO)
