### 2024-02-10 76a1f84
add vignettes user objective function
