+++ inst/R-SWAT/R/objFunctionUI.R
@@ -5,79 +5,91 @@ objFunctionUI <- function(id) {
 
   # shiny::NS(id) returns a namespace function, which was save as `ns` and will
   # invoke later.
-  
+
   ns <- shiny::NS(id)
-  
+
   tagList(
-    
+
     fluidRow(
 
       #-------------------------------------------------------------------------
       # 1. Objective function
-      #-------------------------------------------------------------------------  
+      #-------------------------------------------------------------------------
 
       column(width = 10,
              selectInput("objFunction",
                          width = "40%",
-                         label = "1. Select objective function", 
-                         choices = c('NSE', 'KGE', 'R2', 'RMSE', 'aBIAS', 
+                         label = "1. Select objective function",
+                         choices = c('NSE', 'KGE', 'R2', 'RMSE', 'aBIAS',
                                      'userObjFunction'),
                          multiple = FALSE),
       ),
 
-      conditionalPanel(
-        
-        condition = "input.objFunction == 'userObjFunction'",
-        column(width = 10,
-               selectInput("minOrmax",
-                           width = "40%",
-                           label = "Minimize or maximize the objective function?", 
-                           choices = c('Minimize', 'Maximize'),
-                           multiple = FALSE),
-        ),
-      ),
-      
       column(width = 1,
              tippy::tippy("Help", tooltip = "<span style='font-size:16px;'>
-                   Please select the objective function: NSE = Nash-Sutcliffe 
-                   efficiency, KGE = Kling-Gupta efficiency, R2 = R squared, 
-                   RMSE = Root-mean-square error, aPBIAS =  Absolute Percent bias = 
+                   Please select the objective function: NSE = Nash-Sutcliffe
+                   efficiency, KGE = Kling-Gupta efficiency, R2 = R squared,
+                   RMSE = Root-mean-square error, aPBIAS =  Absolute Percent bias =
                    abs(100 * (sumObserved - sumSimulated))/sumObserved. The same weights
                    are applied for all variables. Users can define their own objective
                    function using the option 'userObjFunction' by modifying the file
                    ./R/userObjFunction.R. See section 5 of R-SWAT wiki page
                    https://github.com/tamnva/R-SWAT/wiki/R-SWAT-User-Manual
-                   <span>", 
-                   allowHTML = TRUE, 
-                   trigger = "click",
-                   theme = "translucent"),
-      ),  
-      
+                   <span>",
+                          allowHTML = TRUE,
+                          trigger = "click",
+                          theme = "translucent"),
+      ),
+
+      conditionalPanel(
+
+        condition = "input.objFunction == 'userObjFunction'",
+        column(width = 10,
+               selectInput("minOrmax",
+                           width = "40%",
+                           label = "Minimize or maximize the objective function?",
+                           choices = c('Minimize', 'Maximize'),
+                           multiple = FALSE),
+        ),
+        column(width = 10,
+               actionButton("getUserObjFunction",
+                            "Load userObjFunction.R",
+                            buttonType = "default",
+                            style="background-color: #87CEFA; border-color: #0d0c0c",
+                            class = NULL),
+        ),
+
+        # Display file name
+        column(width = 10,
+               verbatimTextOutput("userObjFunctionFile", placeholder = TRUE),
+        ),
+      ),
+
       #-------------------------------------------------------------------------
       # 2. Get observed data files
-      #------------------------------------------------------------------------- 
-      
+      #-------------------------------------------------------------------------
+
       column(width = 10,
              HTML("<b>","2. Get observed data files (IMPORTANT CHANGES: Please read 'Help')","</b>"),
       ),
-      
+
       column(width = 1,
              tippy::tippy("Help", tooltip = "<span style='font-size:16px;'>
-                   Load observed data, single file for each variable. 
[…]
