### 2024-02-11 5ed13c2
feat: add more vignettes and update help
