+++ inst/R-SWAT/server.R
@@ -698,7 +698,8 @@ server <- function(input, output, session) {
 
     showModal(modalDialog(
       title = "Help: 4. Select executable SWAT",
-      "Select the executable SWAT or SWAT+ file, for example, swat_32debug.exe",
+      "Select the executable SWAT or SWAT+ file, for example, swat_32debug.exe,
+      you could download these files here https://swat.tamu.edu/software/",
       easyClose = TRUE
     ))
   })
@@ -1195,6 +1196,20 @@ server <- function(input, output, session) {
     # Check if this option is activated
     globalVariable$userReadSwatOutput <<- OutputVar$userReadSwatOutput
 
+    # Display message
+    spsComps::shinyCatch(
+      if (TRUE %in% globalVariable$userReadSwatOutput){
+        output$messageUserReadSwatOutput <- renderText(
+          "MUST check the box below to load 'userReadSwatOutput.R' file")
+        output$userReadSwatOutputFile <- renderText(
+          "Please load 'userReadSwatOutput.R' file")
+      } else {
+        output$messageUserReadSwatOutput <- renderText(" ")
+        output$userReadSwatOutputFile <- renderText(
+          "No input file is needed")
+      },
+      blocking_level = "error")
+
     # Display table of observed file names needed for calibration/optimization
     output$tableOutputExtractionDisplayOnly <- renderDataTable(
       printVariableNameObservedFiles(outputExtraction)
@@ -1209,6 +1224,41 @@ server <- function(input, output, session) {
                         selected = 1),
       blocking_level = "error")
 
+  })
+
+  # ****************************************************************************
+  # Get userObjFunction.R file file
+  # ****************************************************************************
+  observe({
+
+    req(input$getUserReadSwatOutput)
+
+    # Display message
+    spsComps::shinyCatch(
+      if (TRUE %in% globalVariable$userReadSwatOutput){
+        output$userReadSwatOutputFile <- renderText(
+          "Please load 'userReadSwatOutput.R' file")
+      } else {
+        output$userReadSwatOutputFile <- renderText(
+          "No input file is needed")
+      },
+      blocking_level = "error")
+
+
+    # Get full path to userObjFunction.R file
+    shinyjs::disable("getUserReadSwatOutput")
+    spsComps::shinyCatch(globalVariable$getUserReadSwatOutput <<- file.choose(),
+                         blocking_level = "none")
+    shinyjs::enable("getUserReadSwatOutput")
+
+    spsComps::shinyCatch(
+      if (grepl(".R", globalVariable$getUserReadSwatOutput, fixed = TRUE)){
+        output$userReadSwatOutputFile <- renderText(globalVariable$getUserReadSwatOutput)
+        source(globalVariable$getUserReadSwatOutput)
+      } else {
+        output$userReadSwatOutputFile <- renderText("Error: The selected file must be '.R' extention")
+      },
+      blocking_level = "error")
   })
 
   # ****************************************************************************
@@ -1264,6 +1314,19 @@ server <- function(input, output, session) {
     globalVariable$dateRangeCali <<- input$dateRangeCali
   })
 
+  # ****************************************************************************
+  # Help button select file SWAT (or SWAT+) parameter file
+  # ****************************************************************************
+  observe({
+    req(input$helpDateRangeCali)
+
+    showModal(modalDialog(
+      title = "Help: 2. Select date range for calibration",
+      HTML(readLines(file.path(globalVariable$HTMLdir,"HTML",
+                               "helpDateRangeCali.html"),warn=FALSE)),
+      easyClose = TRUE
+    ))
+  })
 
   # ****************************************************************************
   # Help button select file SWAT (or SWAT+) parameter file
@@ -1302,6 +1365,33 @@ server <- function(input, output, session) {
     ))
   })
 
+  # ****************************************************************************
[…]
