+++ man/userReadSwatOutput.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/userReadSwatOutput.R
+\name{userReadSwatOutput}
+\alias{userReadSwatOutput}
+\title{User-defined output extraction}
+\usage{
+userReadSwatOutput()
+}
+\description{
+This is a dummy function. As SWAT/SWAT+ have many outputs, RSWAT only provides
+some standard output extraction. Users want to extract other model outputs can
+implement their own function and replace this function. For more detail see
+RSWAT vignettes (userReadSwatOutput).
+}
+\examples{
+
+# Please see RSWAT vignettes for an example
+
+}
