+++ NAMESPACE
@@ -33,6 +33,8 @@ export(updateCalibrationFile)
 export(updateMultiFile)
 export(updateSingleFile)
 export(updatedFileContent)
+export(userObjFunction)
+export(userReadSwatOutput)
 export(yearlyOutputLoc)
 importFrom(doParallel,registerDoParallel)
 importFrom(foreach,"%dopar%")

[…]
