### 2024-02-14 3e4750e
fix: bugs in readAllOutVar function
