### 2024-02-19 #6320 @diazrenata
@tamnva :wave: I'll be editing this submission! To start, have you had the chance to look at this comment and respond?

> > @tamnva - We'll go ahead and review this. However, note that one of the editors commented
> > > This isn't really packaged to normal R standards (it seems like users are intended to clone the repo and edit scripts before installing the package, which isn't typical) and there's a lot of odd code choices (long chain of "x == y | x == z |", rather than using %in%; bulky wrappers around built-in functions).
>
> Thank you for the Editor's comments. I will check/revise the package regarding those points and provide a detailed response here.
