### 2024-02-19 #6320 @tamnva
> @tamnva 👋 I'll be editing this submission! To start, have you had the chance to look at this comment and respond?
>
> > > @tamnva - We'll go ahead and review this. However, note that one of the editors commented
> > > > This isn't really packaged to normal R standards (it seems like users are intended to clone the repo and edit scripts before installing the package, which isn't typical) and there's a lot of odd code choices (long chain of "x == y | x == z |", rather than using %in%; bulky wrappers around built-in functions).
> >
> >
> > Thank you for the Editor's comments. I will check/revise the package regarding those points and provide a detailed response here.

Yes, I am working on these comments. These are short but major comments. There are still some works needed:

Comment: users are intended to clone the repo and edit scripts before installing the package
Response: There is no need to edit scripts before installing the package now.

Comment: there's a lot of odd code choices (long chain of "x == y | x == z |", rather than using %in%
Response: The updated version uses %in%

Comment: bulky wrappers around built-in functions
Response: I am working to split complex functions into smaller functions (TO DO)
