### 2024-02-20 440b1f3
fix: GUI - turn on/off reminder message
