### 2024-02-21 62bca13
split r functions
