### 2024-02-26 #6320 @diazrenata
Hi @cvitolo and @mps9506, would you be able and willing to review this submission to JOSS?
