### 2024-02-27 b3aaf2d
fix: call SWAT executable file in unix OS platform
