+++ man/updateCalibrationFile.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/updateTxtInOut.R
+% Please edit documentation in R/updateCalibrationFile.R
 \name{updateCalibrationFile}
 \alias{updateCalibrationFile}
 \title{Update calibration.cal file for SWAT+}
