### 2024-03-05 #6320 @diazrenata
Hi @ldecicco-USGS - would you be available and interested in reviewing this submission to JOSS, on an R package for hydrological modeling?
