### 2024-03-06 5e28dda
example: better compression of example data
