### 2024-03-07 33a7f14
fix: note with RCMD check
