+++ R/showRSWAT.R
@@ -1,14 +1,16 @@
-#' 
+#'
 #' Display the Graphical User Interface (GUI) of R-SWAT
-#' 
+#'
 #' @description
 #' Calling this functions will show the GUI of R-SWAT
-#' 
+#'
 #' @return GUI of R-SWAT
 
 #' @examples
 #'
+#'\donttest{
 #' showRSWAT()
+#' }
 #'
 #' @export
 #'
