+++ .github/workflows/R-CMD-check.yaml
@@ -8,6 +8,11 @@ on:
 
 name: R-CMD-check
 
+env:
+  global:
+    - RGL_USE_NULL=TRUE
+    - _R_CHECK_DONTTEST_EXAMPLES_=FALSE
+
 jobs:
   R-CMD-check:
     runs-on: ubuntu-latest
