+++ .github/workflows/R-CMD-check.yaml
@@ -12,10 +12,10 @@ jobs:
   R-CMD-check:
     runs-on: ubuntu-latest
     env:
+      _R_CHECK_DONTTEST_EXAMPLES_: false
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
       R_KEEP_PKG_SOURCE: yes
     steps:
-      - _R_CHECK_DONTTEST_EXAMPLES_=FALSE
       - uses: actions/checkout@v3
 
       - uses: r-lib/actions/setup-r@v2
