+++ NAMESPACE
@@ -43,6 +43,7 @@ importFrom(lhs,randomLHS)
 importFrom(magrittr,"%>%")
 importFrom(parallel,makeCluster)
 importFrom(parallel,stopCluster)
+importFrom(shiny,runApp)
 importFrom(stats,cor)
 importFrom(stats,quantile)
 importFrom(stats,rnorm)
