### 2024-03-10 b64535b
fixed: run SWAT in unix - error in file path
