+++ inst/R-SWAT/server.R
@@ -12,25 +12,6 @@ server <- function(input, output, session) {
   })
 
   #-----------------------------------------------------------------------------
-  # Global variables (use <<- for saving globalVariable inside observe)
-  # These variables are set by default, when opening the tool, will be updated
-  # according the user define in later steps
-  #-----------------------------------------------------------------------------
-  globalVariable <<- list()
-  displayOutput <<- list()
-  displayOutput$plotHru <<- FALSE
-  displayOutput$plotSub <<- FALSE
-  globalVariable$checkSimComplete <<- FALSE
-  globalVariable$loadProject <<- FALSE
-  globalVariable$TxtInOutSWAT <<- FALSE
-  globalVariable$TxtInOutSWATPlus <<- FALSE
-  globalVariable$SWATProject <<- TRUE
-  globalVariable$paraSelection <<- dataParaSelectionSWAT
-  HTMLdir <<- system.file("R-SWAT", package = "RSWAT")
-  globalVariable$nCaliParam <<- nrow(globalVariable$paraSelection)
-  globalVariable$runManualCaliSuccess <<- FALSE
-
-  #-----------------------------------------------------------------------------
   # Global function for running SWAT
   #-----------------------------------------------------------------------------
   SWAT <- function(parameterValue){
@@ -53,10 +34,10 @@ server <- function(input, output, session) {
     colnames(parameterValue) <- NULL
 
     # Save parameter value to the globalVariable
-    globalVariable$parameterValue <<- rbind(globalVariable$parameterValue, parameterValue)
+    globalVariable$parameterValue <-rbind(globalVariable$parameterValue, parameterValue)
 
     # Number of parallel runs cannot be higher than number of input parameter sets
-    globalVariable$ncores <<- min(globalVariable$ncores, nrow(parameterValue))
+    globalVariable$ncores <-min(globalVariable$ncores, nrow(parameterValue))
 
     # Convert input parameter to matrix
     parameterValue <- as.matrix(parameterValue)
@@ -86,19 +67,19 @@ server <- function(input, output, session) {
 
     # If this is a first run
     if (is.null(globalVariable$simData)){
-      globalVariable$simData <<- temp$simData
-      globalVariable$objValueCali <<- temp$objValueCali
-      globalVariable$objValueValid <<- temp$objValueValid
+      globalVariable$simData <-temp$simData
+      globalVariable$objValueCali <-temp$objValueCali
+      globalVariable$objValueValid <- temp$objValueValid
 
     # If this is not the first run, then combine result with the existing data
     } else {
-      globalVariable$simData <<- bindList(globalVariable$simData, temp$simData)
-      globalVariable$objValueCali <<- c(globalVariable$objValueCali, temp$objValueCali)
-      globalVariable$objValueValid <<- c(globalVariable$objValueValid, temp$objValueValid)
+      globalVariable$simData <- bindList(globalVariable$simData, temp$simData)
+      globalVariable$objValueCali <- c(globalVariable$objValueCali, temp$objValueCali)
+      globalVariable$objValueValid <- c(globalVariable$objValueValid, temp$objValueValid)
     }
 
    # Next run, no need to copy unchanged SWAT input files
-   globalVariable$copyUnchangeFiles <<- FALSE
+   globalVariable$copyUnchangeFiles <- FALSE
 
    # Parameter optimization: convert to the minimize problem, regardless of input
    if (globalVariable$minOrmax == "Minimize"){
@@ -157,10 +138,10 @@ server <- function(input, output, session) {
                 nchar(RSWATProjectFile)) == "RSWATproject.rds"){
 
         # Read data in this file
[…]
