+++ inst/R-SWAT/server.R
@@ -11,6 +11,22 @@ server <- function(input, output, session) {
     stopApp()
   })
 
+  # Reactive values
+  displayOutput <- reactiveValues(plotHru = FALSE)
+  displayOutput$plotSub <- FALSE
+
+  globalVariable <- reactiveValues(checkSimComplete = FALSE)
+  globalVariable$loadProject <- FALSE
+  globalVariable$TxtInOutSWAT <- FALSE
+  globalVariable$TxtInOutSWATPlus <- FALSE
+  globalVariable$SWATProject <- TRUE
+  globalVariable$paraSelection <- dataParaSelectionSWAT
+  globalVariable$nCaliParam <- nrow(dataParaSelectionSWAT)
+  globalVariable$runManualCaliSuccess <- FALSE
+  globalVariable$observedData <- list()
+
+  HTMLdir <- reactiveValues(path = system.file("R-SWAT", package = "RSWAT"))
+
   #-----------------------------------------------------------------------------
   # Global function for running SWAT
   #-----------------------------------------------------------------------------
@@ -34,7 +50,7 @@ server <- function(input, output, session) {
     colnames(parameterValue) <- NULL
 
     # Save parameter value to the globalVariable
-    globalVariable$parameterValue <-rbind(globalVariable$parameterValue, parameterValue)
+    globalVariable$parameterValue <- rbind(globalVariable$parameterValue, parameterValue)
 
     # Number of parallel runs cannot be higher than number of input parameter sets
     globalVariable$ncores <-min(globalVariable$ncores, nrow(parameterValue))
@@ -556,14 +572,12 @@ server <- function(input, output, session) {
   # ****************************************************************************
   observe({
     req(input$TxtInOutFolder)
-
     # Check if TxtInOut folder exists
     if(!dir.exists(trimws(input$TxtInOutFolder))){
       output$checkTxtInOutFolder <- renderText("Input folder does not exist")
     } else {
       output$checkTxtInOutFolder <- renderText(" ")
     }
-
     # Check if .hru files exist in this folder
     if (checkDirFileExist(trimws(input$TxtInOutFolder), "", ".cio")){
 
@@ -573,6 +587,7 @@ server <- function(input, output, session) {
 
       # Check if the TxtInOut matches the SWAT project
       if (globalVariable$SWATPlusProject){
+
         if(globalVariable$TxtInOutSWAT){
           output$checkTxtInOutFolder <- renderText(
             "ERROR: This should be TxtInOut of SWAT+"
@@ -720,7 +735,6 @@ server <- function(input, output, session) {
       globalVariable$SWATParam <- loadSwatParam(globalVariable$SWATParamFile)
       output$printSWATParamFile <- renderText(globalVariable$SWATParamFile)
       output$tableSWATParam <- renderDataTable(globalVariable$SWATParam)
-
     } else {
       output$printSWATParamFile <- renderText(
         paste("Error: The selected file must be either 'swatParam.txt'",
@@ -739,7 +753,7 @@ server <- function(input, output, session) {
 
     showModal(modalDialog(
       title = "Help: 5. File with list of SWAT or SWAT+ parameters",
-      HTML(readLines(file.path(HTMLdir,"HTML",
+      HTML(readLines(file.path(HTMLdir$path,"HTML",
 	  "helpSWATparamFile.html"),warn=FALSE)),
       easyClose = TRUE
     ))
@@ -939,7 +953,7 @@ server <- function(input, output, session) {
     req(input$helpParamSelection)
     showModal(modalDialog(
       title = "Help: Parameter Selection",
-	  HTML(readLines(file.path(HTMLdir,"HTML",
[…]
