### 2024-03-13 8c47fbd
check $ not allow with automic vector
