### 2024-04-10 e89d2ff
fix: issue running parallel with userReadSWATOutput
