+++ inst/R-SWAT/server.R
@@ -29,6 +29,9 @@ server <- function(input, output, session) {
   HTMLdir <<- system.file("R-SWAT", package = "RSWAT")
   globalVariable$nCaliParam <<- nrow(globalVariable$paraSelection)
   globalVariable$runManualCaliSuccess <<- FALSE
+  globalVariable$readOutputScript <<- NULL
+  globalVariable$objFunctionScript <<- NULL
+  globalVariable$minOrmax <<- "Maximize"
 
   #-----------------------------------------------------------------------------
   # Global function for running SWAT
@@ -73,7 +76,8 @@ server <- function(input, output, session) {
                globalVariable$copyUnchangeFiles,
                globalVariable$fileCioInfo,
                globalVariable$dateRangeCali,
-               globalVariable$firstRun)
+               globalVariable$firstRun,
+               globalVariable$readOutputScript)
 
     # Objective function with initial parameter set
     temp <- calObjFunction(parameterValue,
@@ -365,7 +369,7 @@ server <- function(input, output, session) {
             }
           })
         }
-		
+
 		# update output extraction reminder text
 		spsComps::shinyCatch(
           if (TRUE %in% globalVariable$userReadSwatOutput){
@@ -379,7 +383,7 @@ server <- function(input, output, session) {
               "No input file is needed")
         },
         blocking_level = "error")
-	  
+
 
         # Show meesage
         showNotification("All project settings were loaded",
@@ -1262,16 +1266,18 @@ server <- function(input, output, session) {
 
     # Get full path to userObjFunction.R file
     shinyjs::disable("getUserReadSwatOutput")
-    spsComps::shinyCatch(globalVariable$getUserReadSwatOutput <<- file.choose(),
+    spsComps::shinyCatch(globalVariable$readOutputScript <<- file.choose(),
                          blocking_level = "none")
     shinyjs::enable("getUserReadSwatOutput")
 
     spsComps::shinyCatch(
-      if (grepl(".R", globalVariable$getUserReadSwatOutput, fixed = TRUE)){
-        output$userReadSwatOutputFile <- renderText(globalVariable$getUserReadSwatOutput)
-        source(globalVariable$getUserReadSwatOutput)
+      if (basename(globalVariable$readOutputScript) == 'userReadSwatOutput.R'){
+        output$userReadSwatOutputFile <- renderText(globalVariable$readOutputScript)
+        source(globalVariable$readOutputScript)
       } else {
-        output$userReadSwatOutputFile <- renderText("Error: The selected file must be '.R' extention")
+        output$userReadSwatOutputFile <- renderText(
+          "Error: The selected file must be 'userReadSwatOutput.R'")
+        globalVariable$readOutputScript <<- NULL
       },
       blocking_level = "error")
   })
@@ -1518,7 +1524,8 @@ server <- function(input, output, session) {
                      copyUnchangeFiles,
                      globalVariable$fileCioInfo,
                      globalVariable$dateRangeCali,
-                     firstRun),
+                     firstRun,
+                     globalVariable$readOutputScript),
           blocking_level = "error"
         )
 
@@ -1553,7 +1560,8 @@ server <- function(input, output, session) {
                        copyUnchangeFiles,
                        globalVariable$fileCioInfo,
                        globalVariable$dateRangeCali,
-                       firstRun),
+                       firstRun,
+                       globalVariable$readOutputScript),
             blocking_level = "error"
           )
 
@@ -1596,7 +1604,9 @@ server <- function(input, output, session) {
             best$perCriteriaValid <- temp$perCriteriaValid
             best$simData <- temp$simData
           } else {
[…]
