### 2024-04-23 #6320 @diazrenata
@jonsampedro, would you be willing and interested in reviewing this submission to JOSS?
