### 2024-05-21 #6320 @jonsampedro
> Hi @jonsampedro - thank you very much! If you would feel able to comment on the code quality and wider relevance of the submission, that would be very much appreciated!

Hi @diazrenata, I understand I should wait to be assigned as a reviewer to start with the review process, right? (Just checking because this is still the pre-review so perhaps other reviewers have been already assigned)
