### 2024-05-21 #6320 @diazrenata
@jonsampedro - yes, it's still as pre-review! If you are OK with serving as
a reviewer, we can go ahead and advance it to the review stage!

On Tue, May 21, 2024 at 1:01 AM jonsampedro ***@***.***>
wrote:

> *External Email*
>
> Hi @jonsampedro <https://github.com/jonsampedro> - thank you very much!
> If you would feel able to comment on the code quality and wider relevance
> of the submission, that would be very much appreciated!
>
> Hi @diazrenata <https://github.com/diazrenata>, I understand I should
> wait to be assigned as a reviewer to start with the review process, right?
> (Just checking because this is still the pre-review so perhaps other
> reviewers have been already assigned)
>
