### 2024-05-28 #6797 @editorialbot
[REVIEW]: RSWAT: An R package for the Soil and Water Assessment Tool models
**Submitting author:** @tamnva (<a href="http://orcid.org/0000-0001-9111-4393">Van Tam Nguyen</a>)
**Repository:** https://github.com/tamnva/R-SWAT
**Branch with paper.md** (empty if default branch): joss_manuscript
**Version:** v4.0.0
