### 2024-06-09 tamnva/R-SWAT#78 @jonsampedro
Code suggestions JOSS
[openjournals/joss-reviews/issues/6797](https://github.com/openjournals/joss-reviews/issues/6797)

The code is clear and the structure seems to be consistent across the different scripts. However, I have tried to run devtools::check() and it does not work due to the following error associated with the vignettes:
![image](https://github.com/tamnva/R-SWAT/assets/78936824/7ad2da60-26a4-4c01-8e7a-3a8481de4413)

Also, I have needed to install some additional dependencies for the check:
![image](https://github.com/tamnva/R-SWAT/assets/78936824/a149c937-3703-493e-82aa-9b5ab693d7f4)

I have some minor comments:
-	Due to the dependency with the “_parallel_” package (which is integrated in core R), RSWAT requires the latest versions of R (>4.1.1), so it needs to be added.
-	There are some other dependencies that are missing and not included in the DESCRIPTION file(e.g., 'magrittr' 2.0.1 is being loaded, but >= 2.0.3 is required)
-	If the selected directory (_tempdir_) does not exist, the package should create the folder in the system
-	Consider creating a documentation website using the “pkgdown” (or a similar) package.
