### 2024-06-09 #6797 @jonsampedro
@diazrenata @tamnva,

I have read the manuscript and installed the software; see specific comments on the manuscript and coding in two issues opened in the package repository (both referenced above).

@tamnva, I have tried to run the package but it is not clear for me if I need additional software (SWAT?) or with the provided example files is enough... If you could provide some additional instructions I would be happy to further test the functionality more in depth and provide additional feedback (if any).

I am happy to clarify any of the raised points or answer any questions if useful.
