### 2024-08-01 af648c8
add SWAT/SWAT+ dependency to README and DESCRIPTION
