+++ DESCRIPTION
@@ -7,7 +7,7 @@ Maintainer: Tam V. Nguyen <tamnva@gmail.com>
 Description: Parameter calibration and uncertainty analysis with SWAT/SWAT+ and their modified versions.
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 7.3.1
+RoxygenNote: 7.3.2
 URL: https://github.com/tamnva/R-SWAT
 BugReports: https://github.com/tamnva/R-SWAT/issues
 Depends: R (>= 4.0)
@@ -40,3 +40,4 @@ Suggests:
     testthat (>= 3.0.0)
 Config/testthat/edition: 3
 VignetteBuilder: knitr
+SystemRequirements: SWAT/SWAT+ executable files (https://swat.tamu.edu/software/)
