### 2024-08-26 #6797 @diazrenata
Hi @tamnva - just checking in to see how things are progressing?
