### 2024-08-30 b303161
feat: show message if cannot find SWAT exe file
