### 2024-08-30 fde30bc
feat: option to extract behavioral parameter sets
