+++ inst/R-SWAT/server.R
@@ -32,6 +32,7 @@ server <- function(input, output, session) {
   globalVariable$readOutputScript <<- NULL
   globalVariable$objFunctionScript <<- NULL
   globalVariable$minOrmax <<- "Maximize"
+  globalVariable$RSWATversion <<- "v4.01"
 
   #-----------------------------------------------------------------------------
   # Global function for running SWAT
@@ -204,7 +205,7 @@ server <- function(input, output, session) {
         output$printSWATParamFile <- renderText(globalVariable$SWATParamFile)
 
         # Update display content of the SWAT parameter file
-        output$tableSWATParam <- renderDataTable(globalVariable$SWATParam)
+        output$tableSWATParam <- renderDT(globalVariable$SWATParam)
 
         #-------------------------------------------------------------------------
         # Update Tab 2: Parameter sampling
@@ -290,7 +291,7 @@ server <- function(input, output, session) {
         }
 
         # Update display corresponding observed file names
-        output$tableOutputExtractionDisplayOnly <- renderDataTable(
+        output$tableOutputExtractionDisplayOnly <- renderDT(
           printVariableNameObservedFiles(globalVariable$outputExtraction))
 
         # Update select date range
@@ -656,7 +657,7 @@ server <- function(input, output, session) {
       }
 
       # Display table of HRU information (land use, soil, slope)
-      output$tableHRUinfo <<- renderDataTable(globalVariable$HRUinfo)
+      output$tableHRUinfo <<- renderDT(globalVariable$HRUinfo)
 
 
     } else {
@@ -742,7 +743,7 @@ server <- function(input, output, session) {
 
       globalVariable$SWATParam <<- loadSwatParam(globalVariable$SWATParamFile)
       output$printSWATParamFile <- renderText(globalVariable$SWATParamFile)
-      output$tableSWATParam <- renderDataTable(globalVariable$SWATParam)
+      output$tableSWATParam <- renderDT(globalVariable$SWATParam)
 
     } else {
       output$printSWATParamFile <- renderText(
@@ -786,7 +787,7 @@ server <- function(input, output, session) {
       # Check if there is no input SWAT parameter file
       if (is.null(globalVariable$SWATParam$parameter)){
         output$tableHelpParameterSelection <-
-          renderDataTable(displayOutput$uniqueHruProperties)
+          renderDT(displayOutput$uniqueHruProperties)
       } else {
 
         SWATParamName <- globalVariable$SWATParam$parameter
@@ -810,7 +811,7 @@ server <- function(input, output, session) {
 
           tempUniqueHruProperties <- cbind(SWATParamName,tempUniqueHruProperties)
 
-          output$tableHelpParameterSelection <- renderDataTable(tempUniqueHruProperties)
+          output$tableHelpParameterSelection <- renderDT(tempUniqueHruProperties)
         }
 
       }
@@ -823,7 +824,7 @@ server <- function(input, output, session) {
                       sep = "")
         # Display SWAT+ help table
         output$tableHelpParameterSelection <-
-          renderDataTable(data.frame(Parameter = temp))
+          renderDT(data.frame(Parameter = temp))
       }
     } else {
       # Display SWAT+ help table
@@ -1230,7 +1231,7 @@ server <- function(input, output, session) {
       blocking_level = "error")
 
     # Display table of observed file names needed for calibration/optimization
-    output$tableOutputExtractionDisplayOnly <- renderDataTable(
+    output$tableOutputExtractionDisplayOnly <- renderDT(
       printVariableNameObservedFiles(outputExtraction)
     )
 
@@ -1901,7 +1902,7 @@ server <- function(input, output, session) {
       colnames(globalVariable$parameterValue) = c("Simulation Nr.",
[…]
