### 2024-09-01 10f03e0
feat: allow extracting results to csv files and behavioral parametersets
