+++ inst/R-SWAT/R/utilities.R
@@ -1,4 +1,3 @@
-
 #-------------------------------------------------------------------------------
 # Binding list object (with 3 levels of list)
 # ------------------------------------------------------------------------------
