### 2024-12-16 052e66a
remove global assignment <<-
