+++ inst/R-SWAT/R/defaultValue.R
@@ -1,244 +1,0 @@
-
-
-# ------------------------------------------------------------------------------
-# Default parameter change values for SWAT
-# ------------------------------------------------------------------------------
-dataParaSelectionSWAT <- data.frame(Parameter = c('GW_DELAY.gw', 'CN2.mgt', 'SOL_K.sol',
-                                             'ALPHA_BF.gw', 'ESCO.hru', 'SURLAG.hru',
-                                             'CH_K2.rte', 'SURLAG.bsn'),
-                               Change = c('absolute', 'relative', 'relative',
-                                          'replace', 'replace', 'replace', 'replace', 'replace'),
-                               Min = c(50,-0.25, -0.25, 0.01, 0.5, 0.1, 0.0, 3.0),
-                               Max = c(450,0.25, 0.25, 0.4, 0.99, 10.0, 0.5, 6.0),
-                               Subbasin = c('All','All', 'All', 'All', 'All', 'All', 'All', NA),
-                               Landuse = c('All','All', 'All', 'All', 'All', 'All', NA, NA),
-                               Soil = c('All','All', 'All', 'All', 'All', 'All', NA, NA),
-                               Slope = c('All', 'All', 'All', 'All', 'All', 'All', NA, NA))
-
-columnsParaSelectionSWAT <- data.frame(title= colnames(dataParaSelectionSWAT),
-                                  source = I(list(NA ,
-                                                  c('absolute', 'relative','replace'),
-                                                  NA,
-                                                  NA,
-                                                  NA,
-                                                  NA,
-                                                  NA,
-                                                  NA)),
-                                  width= c(300,
-                                           300,
-                                           300,
-                                           300,
-                                           300,
-                                           300,
-                                           300,
-                                           300),
-                                  type=c('text',
-                                         'dropdown',
-                                         'numeric',
-                                         'numeric',
-                                         'numeric',
-                                         'text',
-                                         'text',
-                                         'text')
-                                  )
-
-# ------------------------------------------------------------------------------
-# Default parameter change values for SWAT +
-# ------------------------------------------------------------------------------
-dataParaSelectionSWATPlus <- data.frame(Parameter = c('cn2.hru', 'canmx.hru', 'esco.hru',
-                                              'k.sol', 'awc.sol', 'surlag.bsn',
-                                               'msk_co1.bsn', 'msk_co2.bsn', 'chk.rte'
-                                              , 'wd_rto.rte'),
-                                Change = c('relative', 'replace', 'replace',
-                                           'replace', 'replace', 'replace',
-                                           'replace', 'replace', 'replace',
-                                           'replace'),
-                                Min = c(-0.25,1, 0, 0, 0.01, 0.05,0, 0, 3.0, 0.5),
-                                Max = c(0.25,10, 1, 20, 1, 10, 10, 10, 6.0, 20),
-                                Object = c("1:2,3",rep("All",9)),
-                                Conditions = c("hgs=A,B,C; lyr1=1; lyr2=2",
-                                               "landuse = corn, past",
-                                               rep("All",8)))
-
-
-columnsParaSelectionSWATPlus <- data.frame(title= colnames(dataParaSelectionSWATPlus),
-                                   source = I(list(NA ,
-                                                   c('absolute', 'relative','replace'),
-                                                   NA,
-                                                   NA,
-                                                   NA,
-                                                   NA)),
-                                   width= c(200,
-                                            200,
-                                            200,
-                                            200,
-                                            200,
-                                            400
-                                            #200,
-                                            #200,
-                                            #200,
-                                            #200,
-                                            #200,
-                                            #200
-                                            ),
-                                   type=c('text',
-                                          'dropdown',
-                                          'numeric',
-                                          'numeric',
-                                          'text',
-                                          'text')
-)
[…]
