+++ inst/R-SWAT/server.R
@@ -12,29 +12,6 @@ server <- function(input, output, session) {
   })
 
   #-----------------------------------------------------------------------------
-  # Global variables (use <<- for saving globalVariable inside observe)
-  # These variables are set by default, when opening the tool, will be updated
-  # according the user define in later steps
-  #-----------------------------------------------------------------------------
-  globalVariable <<- list()
-  displayOutput <<- list()
-  displayOutput$plotHru <<- FALSE
-  displayOutput$plotSub <<- FALSE
-  globalVariable$checkSimComplete <<- FALSE
-  globalVariable$loadProject <<- FALSE
-  globalVariable$TxtInOutSWAT <<- FALSE
-  globalVariable$TxtInOutSWATPlus <<- FALSE
-  globalVariable$SWATProject <<- TRUE
-  globalVariable$paraSelection <<- dataParaSelectionSWAT
-  HTMLdir <<- system.file("R-SWAT", package = "RSWAT")
-  globalVariable$nCaliParam <<- nrow(globalVariable$paraSelection)
-  globalVariable$runManualCaliSuccess <<- FALSE
-  globalVariable$readOutputScript <<- NULL
-  globalVariable$objFunctionScript <<- NULL
-  globalVariable$minOrmax <<- "Maximize"
-  globalVariable$RSWATversion <<- "v4.01"
-
-  #-----------------------------------------------------------------------------
   # Global function for running SWAT
   #-----------------------------------------------------------------------------
   SWAT <- function(parameterValue){
@@ -57,10 +34,10 @@ server <- function(input, output, session) {
     colnames(parameterValue) <- NULL
 
     # Save parameter value to the globalVariable
-    globalVariable$parameterValue <<- rbind(globalVariable$parameterValue, parameterValue)
+    globalVariable$parameterValue <- rbind(globalVariable$parameterValue, parameterValue)
 
     # Number of parallel runs cannot be higher than number of input parameter sets
-    globalVariable$ncores <<- min(globalVariable$ncores, nrow(parameterValue))
+    globalVariable$ncores <- min(globalVariable$ncores, nrow(parameterValue))
 
     # Convert input parameter to matrix
     parameterValue <- as.matrix(parameterValue)
@@ -91,19 +68,19 @@ server <- function(input, output, session) {
 
     # If this is a first run
     if (is.null(globalVariable$simData)){
-      globalVariable$simData <<- temp$simData
-      globalVariable$objValueCali <<- temp$objValueCali
-      globalVariable$objValueValid <<- temp$objValueValid
+      globalVariable$simData <- temp$simData
+      globalVariable$objValueCali <- temp$objValueCali
+      globalVariable$objValueValid <- temp$objValueValid
 
     # If this is not the first run, then combine result with the existing data
     } else {
-      globalVariable$simData <<- bindList(globalVariable$simData, temp$simData)
-      globalVariable$objValueCali <<- c(globalVariable$objValueCali, temp$objValueCali)
-      globalVariable$objValueValid <<- c(globalVariable$objValueValid, temp$objValueValid)
+      globalVariable$simData <- bindList(globalVariable$simData, temp$simData)
+      globalVariable$objValueCali <- c(globalVariable$objValueCali, temp$objValueCali)
+      globalVariable$objValueValid <- c(globalVariable$objValueValid, temp$objValueValid)
     }
 
    # Next run, no need to copy unchanged SWAT input files
-   globalVariable$copyUnchangeFiles <<- FALSE
+   globalVariable$copyUnchangeFiles <- FALSE
 
    # Parameter optimization: convert to the minimize problem, regardless of input
    if (globalVariable$minOrmax == "Minimize"){
@@ -148,9 +125,7 @@ server <- function(input, output, session) {
   #-----------------------------------------------------------------------------
   # Load project setting
[…]

[…]
