+++ inst/R-SWAT/R/paramSamplingServer.R
@@ -1,0 +1,386 @@
+
+
+paramSamplingServer <- function(id) {
+  moduleServer(id, function(input, output, session) {
+
+    
+    # ****************************************************************************
+    # Display help for parameter selection
+    # ****************************************************************************
+    
+    observeEvent(input$helpParameterSelection, {
+      print("ok14")
+      # This first if command prevents the app crashed when no input is given
+      if (is.data.frame(globalVariable$SWATParam)){
+        # Check if there is no input SWAT parameter file
+        if (is.null(globalVariable$SWATParam$parameter)){
+          output$tableHelpParameterSelection <-
+            renderDT(displayOutput$uniqueHruProperties)
+        } else {
+          
+          SWATParamName <- globalVariable$SWATParam$parameter
+          nSWATParamName <- length(SWATParamName)
+          
+          if(is.null(displayOutput$uniqueHruProperties)){
+            output$tableHelpParameterSelection <- NULL
+          } else {
+            nRowUniqueHRU <- nrow(displayOutput$uniqueHruProperties)
+            compareLength <- max(nRowUniqueHRU, nSWATParamName)
+            if (nRowUniqueHRU < compareLength){
+              newRow <- data.frame(matrix(rep(NA, (compareLength - nRowUniqueHRU)*4),
+                                          ncol = 4))
+              names(newRow) <- names(displayOutput$uniqueHruProperties)
+              tempUniqueHruProperties <- rbind(displayOutput$uniqueHruProperties,
+                                               newRow)
+              
+            } else {
+              SWATParamName <- c(SWATParamName, rep(NA, compareLength - nSWATParamName))
+            }
+            
+            tempUniqueHruProperties <- cbind(SWATParamName,tempUniqueHruProperties)
+            
+            output$tableHelpParameterSelection <- renderDT(tempUniqueHruProperties)
+          }
+          
+        }
+        
+        # If this is a SWAT+ project
+        if (ncol(globalVariable$SWATParam) != 7){
+          
+          # Get SWAT+ parameter name
+          temp <- paste(globalVariable$SWATParam[, 1], ".", globalVariable$SWATParam[, 2],
+                        sep = "")
+          # Display SWAT+ help table
+          output$tableHelpParameterSelection <-
+            renderDT(data.frame(Parameter = temp))
+        }
+      } else {
+        # Display SWAT+ help table
+        output$tableHelpParameterSelection <- NULL
+      }
+      print("ok15")
+    })
+    print("ok14-1")
+    # ****************************************************************************
+    # Help parameter selection
+    # ****************************************************************************
+    observeEvent(input$helpParam, {
+      print("ok14-1")
+      showModal(modalDialog(
+        title = "Help: 1. Display help for parameter selection",
+        "If you don't know the parameter, subbasin, land use, and slope names,
+      check this box. We will get this information from the TxtInOut folder for you",
+        easyClose = TRUE
+      ))
+    })
+    
+    # ****************************************************************************
+    # Check input 'Select SWAT parameters for calibration'
+    # ****************************************************************************
+    observeEvent(input$checkParameterTableButton, {
+      print("ok16")
+      spsComps::shinyCatch(
+        checkParameterTable <- checkSwatParameterName(globalVariable$paraSelection,
+                                                      globalVariable$SWATParam,
+                                                      globalVariable$HRUinfo,
+                                                      globalVariable$SWATProject),
+        blocking_level = "error"
+      )
+      
+      output$checkParameterTableTxtOuput <- renderText(checkParameterTable$checkMessage)
+    })
+    
+    # ****************************************************************************
+    # Save "SWAT parameters for calibration" to global variable
[…]
