+++ inst/R-SWAT/R/generalSettingServer.R
@@ -1,0 +1,362 @@
+
+generalSettingServer <- function(id) {
+  moduleServer(id, function(input, output, session) {
+    
+    #-----------------------------------------------------------------------------
+    # Tab 1. General Setting
+    #-----------------------------------------------------------------------------
+    # ****************************************************************************
+    # Check SWAT or SWAT+ project
+    # ****************************************************************************
+    observeEvent(input$SWATorSWATplus, {
+      
+      # Check SWAT or SWAT+ project
+      if (input$SWATorSWATplus == "SWAT"){
+        globalVariable$SWATProject <- TRUE
+        globalVariable$SWATPlusProject <- FALSE
+      } else {
+        globalVariable$SWATProject <- FALSE
+        globalVariable$SWATPlusProject <- TRUE
+      }
+      
+      # Initially there is no warning message
+      output$checkTxtInOutFolder <- renderText(" ")
+      
+      # Check if the TxtInOut matches the SWAT project
+      if (globalVariable$SWATPlusProject){
+        if(globalVariable$TxtInOutSWAT){
+          output$checkTxtInOutFolder <- renderText(
+            "ERROR: This should be TxtInOut of SWAT+"
+          )
+        }
+      } else {
+        if(globalVariable$TxtInOutSWATPlus){
+          output$checkTxtInOutFolder <- renderText(
+            "ERROR: This should be TxtInOut of SWAT"
+          )
+        }
+      }
+      
+      # ****************************************************************************
+      # Select SWAT parameters for calibration and/or sensitivity: Default setting
+      # ****************************************************************************
+      print("ok1")
+      if (!globalVariable$loadProject){
+        
+        # Update template for parameter and output extraction according to the SWAT project
+        if (globalVariable$SWATProject){
+          print("ok2")
+          # Example of parameter selection for SWAT project
+          output$tableParaSelection <-
+            excelR::renderExcel(excelR::excelTable(data = dataParaSelectionSWAT,
+                                                   columns = columnsParaSelectionSWAT,
+                                                   editable = TRUE,
+                                                   allowInsertRow = TRUE,
+                                                   allowInsertColumn = TRUE,
+                                                   allowDeleteColumn = TRUE,
+                                                   allowDeleteRow = TRUE,
+                                                   rowDrag = TRUE,
+                                                   columnResize = FALSE,
+                                                   wordWrap = TRUE))
+          
+          # Example of output extraction for SWAT project
+          output$tableOutputExtraction <-
+            excelR::renderExcel(excelR::excelTable(data = dataOutputExtractionSWAT,
+                                                   columns = columnsOutputExtractionSWAT,
+                                                   editable = TRUE,
+                                                   allowInsertRow = TRUE,
+                                                   allowInsertColumn = FALSE,
+                                                   allowDeleteColumn = FALSE,
+                                                   allowDeleteRow = TRUE,
+                                                   rowDrag = FALSE,
+                                                   columnResize = FALSE,
+                                                   wordWrap = TRUE))
+          
+          # If this is SWAT+ project
+        } else {
+          
+          # Example of parameter selection for SWAT+ project
+          output$tableParaSelection <-
+            excelR::renderExcel(excelR::excelTable(data = dataParaSelectionSWATPlus,
+                                                   columns = columnsParaSelectionSWATPlus,
[…]

[…]
