### 2024-12-18 2c7402b
update name sapce id (does not work with ns = NS(id))
