+++ inst/R-SWAT/R/generalSettingUI.R
@@ -3,91 +3,89 @@
 
 generalSettingUI <- function(id) {
 
-  # shiny::NS(id) returns a namespace function, which was save as `ns` and will
-  # invoke later.
-  
-  ns <- shiny::NS(id)
-  
   tagList(
-    
-    fluidRow( 
-      
-      #-------------------------------------------------------------------------
-      # 1. SWAT or SWAT+ project
-      #-------------------------------------------------------------------------
+
+    fluidRow(
+      #------------------------------------------------------------------------#
+      # 1. SWAT or SWAT+ project                                               #
+      #------------------------------------------------------------------------#
       column(width = 10,
-             selectInput("SWATorSWATplus",
+             selectInput(shiny::NS(id, "SWATorSWATplus"),
                          width = "50%",
-                         label = "1. SWAT or SWAT+ project", 
+                         label = "1. SWAT or SWAT+ project",
                          choices = c('SWAT', 'SWAT+'),
                          multiple = FALSE),
       ),
-      
+
       column(width = 1,
-             actionButton("helpSWATorSWATplus", 
-                          "Help",
-                          buttonType = "default",
-                          style="background-color: none; border-color: none",
-                          class = NULL),
-      ),
-      
-      #-------------------------------------------------------------------------
-      # 2. Working folder
-      #-------------------------------------------------------------------------
-      column(width = 10,
-             textInput("workingFolder",
-                       "2. Working folder",
-                       "C:/RSWAT_demo/workingFolder",
-                       width = "100%"
-             ),
-             span(textOutput("checkWorkingFolder"), style="color:red"),
-      ),
-      
-      column(width = 1,
-             actionButton("helpworkingFolder", 
+             actionButton(shiny::NS(id, "helpSWATorSWATplus"),
                           "Help",
                           buttonType = "default",
                           style="background-color: none; border-color: none",
                           class = NULL),
       ),
 
-      #-------------------------------------------------------------------------
-      # 3. TxtInOut folder
-      #-------------------------------------------------------------------------
+      #------------------------------------------------------------------------#
+      # 2. Working folder                                                      #
+      #------------------------------------------------------------------------#
       column(width = 10,
-             textInput("TxtInOutFolder",
-                       "3. TxtInOut folder",
-                       "C:/RSWAT_demo/R-SWAT-master/data/TxtInOut",
+             textInput(shiny::NS(id, "workingFolder"),
+                       "2. Working folder",
+                       "C:/RSWAT_demo/workingFolder",
                        width = "100%"
              ),
-             span(textOutput("checkTxtInOutFolder"), style="color:red"),
-             
+
+             span(textOutput(shiny::NS(id, "checkWorkingFolder")),
+                  style="color:red"),
       ),
-      
+
       column(width = 1,
-             actionButton("helpTxtInOutFolder", 
+             actionButton(shiny::NS(id, "helpworkingFolder"),
                           "Help",
                           buttonType = "default",
                           style="background-color: none; border-color: none",
                           class = NULL),
-            
       ),
-      
+
+      #------------------------------------------------------------------------#
+      # 3. TxtInOut folder                                                     #
+      #------------------------------------------------------------------------#
       column(width = 10,
-             checkboxInput('checkTxtInOutFolder', 'Display HRU info from TxtInOut 
-                           folder', value = FALSE, width = NULL),
+             textInput(shiny::NS(id, "TxtInOutFolder"),
+                       "3. TxtInOut folder",
+                       "C:/RSWAT_demo/R-SWAT-master/data/TxtInOut",
+                       width = "100%"
+             ),
[…]
