+++ inst/R-SWAT/R/generalSettingServer.R
@@ -1,7 +1,7 @@
 
 generalSettingServer <- function(id) {
   moduleServer(id, function(input, output, session) {
-    
+
     #-----------------------------------------------------------------------------
     # Tab 1. General Setting
     #-----------------------------------------------------------------------------
@@ -9,7 +9,7 @@ generalSettingServer <- function(id) {
     # Check SWAT or SWAT+ project
     # ****************************************************************************
     observeEvent(input$SWATorSWATplus, {
-      
+
       # Check SWAT or SWAT+ project
       if (input$SWATorSWATplus == "SWAT"){
         globalVariable$SWATProject <- TRUE
@@ -18,10 +18,11 @@ generalSettingServer <- function(id) {
         globalVariable$SWATProject <- FALSE
         globalVariable$SWATPlusProject <- TRUE
       }
-      
+      print("okcheck")
       # Initially there is no warning message
-      output$checkTxtInOutFolder <- renderText(" ")
-      
+      output$checkTxtInOutFolder <- renderText("546546455")
+      print("okcheck2")
+
       # Check if the TxtInOut matches the SWAT project
       if (globalVariable$SWATPlusProject){
         if(globalVariable$TxtInOutSWAT){
@@ -36,13 +37,13 @@ generalSettingServer <- function(id) {
           )
         }
       }
-      
+
       # ****************************************************************************
       # Select SWAT parameters for calibration and/or sensitivity: Default setting
       # ****************************************************************************
       print("ok1")
       if (!globalVariable$loadProject){
-        
+
         # Update template for parameter and output extraction according to the SWAT project
         if (globalVariable$SWATProject){
           print("ok2")
@@ -58,7 +59,7 @@ generalSettingServer <- function(id) {
                                                    rowDrag = TRUE,
                                                    columnResize = FALSE,
                                                    wordWrap = TRUE))
-          
+
           # Example of output extraction for SWAT project
           output$tableOutputExtraction <-
             excelR::renderExcel(excelR::excelTable(data = dataOutputExtractionSWAT,
@@ -71,10 +72,10 @@ generalSettingServer <- function(id) {
                                                    rowDrag = FALSE,
                                                    columnResize = FALSE,
                                                    wordWrap = TRUE))
-          
+
           # If this is SWAT+ project
         } else {
-          
+
           # Example of parameter selection for SWAT+ project
           output$tableParaSelection <-
             excelR::renderExcel(excelR::excelTable(data = dataParaSelectionSWATPlus,
@@ -87,7 +88,7 @@ generalSettingServer <- function(id) {
                                                    rowDrag = TRUE,
                                                    columnResize = FALSE,
                                                    wordWrap = TRUE))
-          
+
           # Example of output extraction for SWAT+ project
           output$tableOutputExtraction <-
             excelR::renderExcel(excelR::excelTable(data = dataOutputExtractionSWATPlus,
@@ -100,39 +101,40 @@ generalSettingServer <- function(id) {
                                                    rowDrag = FALSE,
                                                    columnResize = FALSE,
                                                    wordWrap = TRUE))
-          
+
           globalVariable$paraSelection <- dataParaSelectionSWATPlus
         }
         print("ok3")
       }
       print("ok4")
     })
-    
+
     # ****************************************************************************
     # Help button select SWAT or SWAT+ project
     # ****************************************************************************
[…]
