+++ inst/R-SWAT/server.R
@@ -93,37 +93,6 @@ server <- function(input, output, session) {
    return(output)
   }
 
-
-
-  #-----------------------------------------------------------------------------
-  # Tab 1. Parameter sampling
-  #-----------------------------------------------------------------------------
-  generalSettingServer("general_setting_id")
-
-
-
-  #-----------------------------------------------------------------------------
-  # Tab 2. Parameter sampling
-  #-----------------------------------------------------------------------------
-  paramSamplingServer("param_sampling_id")
-
-  #-----------------------------------------------------------------------------
-  # Tab 3. Run SWAT
-  #-----------------------------------------------------------------------------
-  runSWATServer("run_swat_id")
-
-  #-----------------------------------------------------------------------------
-  # Tab 4. Evaluate output
-  #-----------------------------------------------------------------------------
-  objFunctionServer("obj_function_id")
-  sensAnalysisServer("sens_analysis_id")
-  paraOptUncerServer("para_opt_uncer_id")
-  # ****************************************************************************
-  # 5. R-SWAT Education
-  # ****************************************************************************
-  swatEduServer("swat_edu_id")
-  #-----------------------------------------------------------------------------
-
   #-----------------------------------------------------------------------------
   # Save project setting
   #-----------------------------------------------------------------------------
@@ -415,4 +384,30 @@ server <- function(input, output, session) {
       blocking_level = "error")
   })
 
+  #-----------------------------------------------------------------------------
+  # Tab 1. Parameter sampling
+  #-----------------------------------------------------------------------------
+  generalSettingServer("general_setting_id")
+
+  #-----------------------------------------------------------------------------
+  # Tab 2. Parameter sampling
+  #-----------------------------------------------------------------------------
+  paramSamplingServer("param_sampling_id")
+
+  #-----------------------------------------------------------------------------
+  # Tab 3. Run SWAT
+  #-----------------------------------------------------------------------------
+  runSWATServer("run_swat_id")
+
+  #-----------------------------------------------------------------------------
+  # Tab 4. Evaluate output
+  #-----------------------------------------------------------------------------
+  objFunctionServer("obj_function_id")
+  sensAnalysisServer("sens_analysis_id")
+  paraOptUncerServer("para_opt_uncer_id")
+  # ****************************************************************************
+  # 5. R-SWAT Education
+  # ****************************************************************************
+  swatEduServer("swat_edu_id")
+  #-----------------------------------------------------------------------------
 }
