### 2024-12-20 6f90d7c
finish update general setting
