+++ inst/R-SWAT/R/generalSettingUI.R
@@ -4,7 +4,6 @@
 generalSettingUI <- function(id) {
 
   tagList(
-
     fluidRow(
       #------------------------------------------------------------------------#
       # 1. SWAT or SWAT+ project                                               #
@@ -16,7 +15,6 @@ generalSettingUI <- function(id) {
                          choices = c('SWAT', 'SWAT+'),
                          multiple = FALSE),
       ),
-
       column(width = 1,
              actionButton(shiny::NS(id, "helpSWATorSWATplus"),
                           "Help",
@@ -34,11 +32,9 @@ generalSettingUI <- function(id) {
                        "C:/RSWAT_demo/workingFolder",
                        width = "100%"
              ),
-
              span(textOutput(shiny::NS(id, "checkWorkingFolder")),
                   style="color:red"),
       ),
-
       column(width = 1,
              actionButton(shiny::NS(id, "helpworkingFolder"),
                           "Help",
@@ -56,25 +52,20 @@ generalSettingUI <- function(id) {
                        "C:/RSWAT_demo/R-SWAT-master/data/TxtInOut",
                        width = "100%"
              ),
-
-             span(textOutput("checkTxtInOutFolder"), style="color:red"),
+             span(textOutput(shiny::NS(id,"checkTxtInOutFolder")), style="color:red"),
       ),
-
       column(width = 1,
              actionButton(shiny::NS(id, "helpTxtInOutFolder"),
                           "Help",
                           buttonType = "default",
                           style="background-color: none; border-color: none",
                           class = NULL),
-
       ),
-
       column(width = 10,
              checkboxInput(shiny::NS(id, 'checkTxtInOutFolder'),
-             'Display HRU info from TxtInOut folder',
-             value = FALSE, width = NULL),
+                           'Display HRU info from TxtInOut folder',
+                           value = FALSE, width = NULL),
       ),
-
       conditionalPanel(
         condition = "input.checkTxtInOutFolder == 1",
         column(width = 10,
@@ -96,7 +87,6 @@ generalSettingUI <- function(id) {
         ),
         ns = shiny::NS(id)
       ),
-
       # If SWAT+ project (step 1) is selected
       conditionalPanel(
         condition = "input.SWATorSWATplus == 'SWAT+'",
@@ -107,7 +97,6 @@ generalSettingUI <- function(id) {
         ),
         ns = shiny::NS(id)
       ),
-
       column(width = 1,
              actionButton(shiny::NS(id, "helpgetSWATexe"),
                           "Help",
@@ -115,7 +104,6 @@ generalSettingUI <- function(id) {
                           style="background-color: none; border-color: none",
                           class = NULL),
       ),
-
       # Pop up window for selecting the executable SWAT or SWAT+
       column(width = 10,
              actionButton(shiny::NS(id, "getSWATexe"),
@@ -123,16 +111,14 @@ generalSettingUI <- function(id) {
                           buttonType = "default",
                           style="background-color: #87CEFA; border-color: #0d0c0c",
                           class = NULL),
-             verbatimTextOutput("printSWATexe", placeholder = TRUE),
+             verbatimTextOutput(shiny::NS(id,"printSWATexe"), placeholder = TRUE),
       ),
-
 
       #------------------------------------------------------------------------#
       # 5. Files with list of all SWAT parameters                              #
       #------------------------------------------------------------------------#
       # Adjust the title of this section depending on SWAT or SWAT+ project
       div( style = "margin-top: 15em",
-
            # If this is a SWAT project
            conditionalPanel(
              condition = "input.SWATorSWATplus == 'SWAT'",
@@ -143,7 +129,6 @@ generalSettingUI <- function(id) {
              ),
              ns = shiny::NS(id)
            ),
-
            # If this is a SWAT+ project
[…]
