+++ inst/R-SWAT/R/generalSettingServer.R
@@ -1,13 +1,11 @@
 
 generalSettingServer <- function(id) {
+
   moduleServer(id, function(input, output, session) {
 
-    #-----------------------------------------------------------------------------
-    # Tab 1. General Setting
-    #-----------------------------------------------------------------------------
-    # ****************************************************************************
-    # Check SWAT or SWAT+ project
-    # ****************************************************************************
+    #--------------------------------------------------------------------------#
+    #                      Check SWAT or SWAT+ project                         #
+    #--------------------------------------------------------------------------#
     observeEvent(input$SWATorSWATplus, {
 
       # Check SWAT or SWAT+ project
@@ -18,95 +16,84 @@ generalSettingServer <- function(id) {
         globalVariable$SWATProject <- FALSE
         globalVariable$SWATPlusProject <- TRUE
       }
-      print("okcheck")
-      # Initially there is no warning message
-      output$checkTxtInOutFolder <- renderText("546546455")
-      print("okcheck2")
 
-      # Check if the TxtInOut matches the SWAT project
-      if (globalVariable$SWATPlusProject){
-        if(globalVariable$TxtInOutSWAT){
-          output$checkTxtInOutFolder <- renderText(
-            "ERROR: This should be TxtInOut of SWAT+"
-          )
-        }
-      } else {
-        if(globalVariable$TxtInOutSWATPlus){
-          output$checkTxtInOutFolder <- renderText(
-            "ERROR: This should be TxtInOut of SWAT"
-          )
-        }
-      }
+      out_check <- check_txtinout_message(input$TxtInOutFolder,
+                                          globalVariable$SWATPlusProject)
+      globalVariable$TxtInOutSWAT <- out_check$txtinout_swat
+      globalVariable$TxtInOutSWATPlus <- out_check$txtinout_swat_plus
+      output$checkTxtInOutFolder <- renderText(out_check$out_message)
 
       # ****************************************************************************
       # Select SWAT parameters for calibration and/or sensitivity: Default setting
       # ****************************************************************************
-      print("ok1")
+
       if (!globalVariable$loadProject){
 
         # Update template for parameter and output extraction according to the SWAT project
         if (globalVariable$SWATProject){
-          print("ok2")
+
           # Example of parameter selection for SWAT project
           output$tableParaSelection <-
-            excelR::renderExcel(excelR::excelTable(data = dataParaSelectionSWAT,
-                                                   columns = columnsParaSelectionSWAT,
-                                                   editable = TRUE,
-                                                   allowInsertRow = TRUE,
-                                                   allowInsertColumn = TRUE,
-                                                   allowDeleteColumn = TRUE,
-                                                   allowDeleteRow = TRUE,
-                                                   rowDrag = TRUE,
-                                                   columnResize = FALSE,
-                                                   wordWrap = TRUE))
+            excelR::renderExcel(excelR::excelTable(
+              data = dataParaSelectionSWAT,
+              columns = columnsParaSelectionSWAT,
+              editable = TRUE,
+              allowInsertRow = TRUE,
+              allowInsertColumn = TRUE,
+              allowDeleteColumn = TRUE,
+              allowDeleteRow = TRUE,
+              rowDrag = TRUE,
+              columnResize = FALSE,
+              wordWrap = TRUE))
 
           # Example of output extraction for SWAT project
           output$tableOutputExtraction <-
-            excelR::renderExcel(excelR::excelTable(data = dataOutputExtractionSWAT,
-                                                   columns = columnsOutputExtractionSWAT,
-                                                   editable = TRUE,
[…]
