+++ inst/R-SWAT/R/paramSamplingUI.R
@@ -1,129 +1,111 @@
 
-# Graphical interface for parameter sampling
-
 paramSamplingUI <- function(id) {
 
-  # shiny::NS(id) returns a namespace function, which was saved as `ns` and will
-  # invoke later.
-  
-  ns <- shiny::NS(id)
-  
   tagList(
-    
     fluidRow(
 
-      #-------------------------------------------------------------------------
-      # 1. Parameter selection table
-      #-------------------------------------------------------------------------
+      #------------------------------------------------------------------------#
+      # 1. Parameter selection table                                           #
+      #------------------------------------------------------------------------#
       column(width = 10,
-             HTML("<b>","1. Select SWAT parameters for calibration and/or 
+             HTML("<b>","1. Select SWAT parameters for calibration and/or
                   sensitivity analysis","</b>"),
       ),
-  
       column(width = 10,
-             checkboxInput('helpParameterSelection', 
-                           'Display help for parameter selection', 
+             checkboxInput(shiny::NS(id,'helpParameterSelection'),
+                           'Display help for parameter selection',
                            value = FALSE, width = NULL),
       ),
-      
       column(width = 1,
-             actionButton("helpParam", 
+             actionButton(shiny::NS(id,"helpParam"),
                           "Help",
                           buttonType = "default",
                           style="background-color: none; border-color: none",
                           class = NULL),
-             
       ),
-      
+
       conditionalPanel(
         condition = "input.helpParameterSelection == 1",
         column(width = 10,
-               dataTableOutput('tableHelpParameterSelection'),
+               dataTableOutput(shiny::NS(id,'tableHelpParameterSelection')),
         ),
+        ns = shiny::NS(id)
       ),
-      
       column(width = 10,
-             excelR::excelOutput("tableParaSelection", 
-                         width = "100%", 
-                         height = "1px"),
+             excelR::excelOutput(shiny::NS(id, "tableParaSelection"),
+                         width = "100%",
+                         height = "0px"),
       ),
-      
       column(width = 1,
-             actionButton("helpParamSelection", 
+             actionButton(shiny::NS(id,"helpParamSelection"),
                           "Help",
                           buttonType = "default",
                           style="background-color: none; border-color: none",
                           class = NULL),
       ),
-      
-      
       column(width = 10,
-             actionButton("checkParameterTableButton", 
+             actionButton(shiny::NS(id,"checkParameterTableButton"),
                           "Click here to check the input table (IMPORTANT)",
-                          #icon("paper-plane"), 
-                          style="background-color: #fa8787; 
+                          #icon("paper-plane"),
+                          style="background-color: #fa8787;
                           border-color: #2e6da4"),
              verbatimTextOutput('checkParameterTableTxtOuput'),
       ),
-      
-      div( style = "margin-bottom: 10em",  
+      div( style = "margin-bottom: 10em",
            column(width = 10,
            ),
       ),
 
-      #-------------------------------------------------------------------------
-      # 2. Select sensitivity or calibration approach
-      #-------------------------------------------------------------------------
+      #------------------------------------------------------------------------#
+      # 2. Select sensitivity or calibration approach                          #
+      #------------------------------------------------------------------------#
       column(width = 10,
-             selectInput("samplingApproach",
-                         width = "50%",
-                         label = "2. Select sensitivity or calibration approach", 
-                         choices = c('Sensi_Cali_(uniform_Latin_Hypercube_Sampling)', 
-                                     'Sensi_(from_sensitivity_package)',
-                                     'Sensi_(from_userDefined_package)',
-                                     'Cali_(from_optimization_package)',
-                                     'Cali_(from_nloptr_package)',
[…]
