### 2024-12-20 a3aaac7
redo general setting ok
