### 2025-02-21 #6797 @tamnva
Hi @diazrenata and @Kevin-Mattheus-Moerman,

Apologies for my late response. The workload turned out to be more than I had anticipated. I have been working on incorporating the reviewers' comments ([in this branch](https://github.com/tamnva/R-SWAT/tree/code_cleaning)), but I encountered some challenges that exceed my current R knowledge.

Due to these difficulties, I would like to withdraw this submission.

That said, I am still working on improving the software based on the feedback from @mps9506 and @jonsampedro.

Thank you for your feedback, and I appreciate your understanding.

Best,

Tam
